requirements.txt
robotframework          == 6.0.1
strict-rfc3339          == 0.7
requests                == 2.26.0
decorator               == 5.1.0

